Style: Fix animation delays
